feat: create a generator module
