+++ multiauth/providers/manual.py
@@ -31,40 +31,3 @@ def manual_authenticator(user: User) -> AuthResponse:
     auth_response['tech'] = AuthTech.MANUAL
 
     return auth_response
-
-
-def serialize_headers_to_manual(headers: dict[str, str] | list[str] | str) -> tuple[dict, dict]:
-    """Serialize raw headers in "manual" auth format."""
-
-    headers_dict: dict[str, str] = {}
-
-    if isinstance(headers, str):
-        headers = [headers]
-
-    if isinstance(headers, list):
-
-        for header in headers:
-            header_split = header.split(':')
-            header_name = header_split[0].strip()
-            header_value = ':'.join(header_split[1:]).strip()
-            headers_dict[header_name] = header_value
-
-    elif isinstance(headers, dict):
-        headers_dict = headers
-
-    auth_name = 'manual_headers'
-
-    auths: dict = {
-        auth_name: {
-            'tech': AuthTech.MANUAL,
-        },
-    }
-
-    users: dict = {
-        'manual_user': {
-            'headers': headers_dict,
-            'auth': auth_name
-        },
-    }
-
-    return auths, users
