+++ tests/providers/test_manual_auth.py
@@ -5,7 +5,7 @@
 import pytest
 
 from multiauth import MultiAuth
-from multiauth.providers.manual import manual_authenticator, serialize_headers_to_manual
+from multiauth.providers.manual import manual_authenticator
 
 
 @pytest.fixture
@@ -83,26 +83,3 @@ def test_manual_handler_headers(users_one_header: dict, auth: dict) -> None:
     auth_response = manual_authenticator(MultiAuth.serialize_users(auth, users_one_header)['manual_user'])
 
     assert auth_response['headers']['Authorization'] == 'Bearer 12345'
-
-
-def test_serialize_headers(auth: dict, users_one_header: dict, users_two_headers: dict) -> None:
-    """Test serialize_headers."""
-
-    headers_str = 'Authorization: Bearer 12345'
-    headers_list = ['Authorization: Bearer 12345', 'Content-Type: application/json']
-    headers_dict = {'Authorization': 'Bearer 12345', 'Content-Type': 'application/json'}
-
-    auths_str, users_str = serialize_headers_to_manual(headers_str)
-
-    assert auths_str == auth
-    assert users_str == users_one_header
-
-    auths_list, users_list = serialize_headers_to_manual(headers_list)
-
-    assert auths_list == auth
-    assert users_list == users_two_headers
-
-    auths_dict, users_dict = serialize_headers_to_manual(headers_dict)
-
-    assert auths_dict == auth
-    assert users_dict == users_two_headers
